// shared/enums/rashi.ts
/**
 * KarmicLens
 * Rashi Enumeration
 *
 * Version: 1.0.0
 *
 * Defines the twelve Rashis of the sidereal
 * Jyotisha zodiac in their traditional sequence.
 */

export enum Rashi {
    Mesha = "Mesha",
    Vrishabha = "Vrishabha",
    Mithuna = "Mithuna",
    Karka = "Karka",
    Simha = "Simha",
    Kanya = "Kanya",
    Tula = "Tula",
    Vrischika = "Vrischika",
    Dhanu = "Dhanu",
    Makara = "Makara",
    Kumbha = "Kumbha",
    Meena = "Meena",
}